Handle ibm watson import error, convert watson tests to pyttest (#1649)
