+++ setup.cfg
@@ -92,9 +92,8 @@ connector_webex =
 # parsers
 parser_dialogflow =
 	dialogflow>=0.8.0
-# TODO: Add this module back in when it can handle ImportError as the dialogflow module does
-#parser_watson =
-#	ibm-watson>=4.4.1
+parser_watson =
+	ibm-watson>=4.4.1
 # databases
 database_redis =
 	aioredis>=1.3.1
